支付宝将加签转为服务器端
1、支付宝将加签转为服务器端
2、修复订单显示的时间时区问题

+++ ECMobileFramework/ECMobileFramework/framework/mvc/controller/Bee_Message.mm
@@ -236,7 +236,7 @@ static NSMutableArray *			__globalExecuters = nil;
 
 - (NSString *)description
 {
-	return [NSString stringWithFormat:@"msg '%@', state %u", self.message, self.state];
+	return [NSString stringWithFormat:@"msg '%@', state %ld", self.message, (long)self.state];
 }
 
 - (void)dealloc
